requirements/local.txt
# Local development dependencies go here
-r base.txt
-r test.txt

# Debugging Tools
django-debug-toolbar==3.1.1
django-extensions==3.0.9
